Switches dietary picker, build modal, submit to routes to collection page, clears form
