+++ app/(tabs)/(collections)/index.jsx
@@ -1,4 +1,4 @@
-import { CollectionList } from '@/components/CollectionList';
+import { CollectionList } from "@/components/CollectionList";
 import { Link, useLocalSearchParams } from "expo-router";
 import React, { useEffect, useState } from "react";
 import {
@@ -12,7 +12,6 @@ import {
 import firestore from "@react-native-firebase/firestore";
 import auth from "@react-native-firebase/auth";
 import { firebase } from "@react-native-firebase/auth";
-
 
 const styles = StyleSheet.create({
   container: {
@@ -73,12 +72,13 @@ const styles = StyleSheet.create({
 });
 
 export default function HomeScreen() {
-  const {collectionName, url, collectionDescription, image, collectionAdded} = useLocalSearchParams()
+  const { collectionName, url, collectionDescription, image, collectionAdded } =
+    useLocalSearchParams();
   const [collections, setCollections] = useState([]);
   const [loading, setLoading] = useState(false);
   const [empty, setEmpty] = useState(false);
   const user = firebase.auth().currentUser;
-  
+
   useEffect(() => {
     setLoading(true);
     setEmpty(false);
@@ -102,11 +102,10 @@ export default function HomeScreen() {
           setEmpty(true);
         }
 
-
         setCollections(filteredList);
         setLoading(false);
       });
-  }, [ collectionName, url, collectionDescription, image, collectionAdded]);
+  }, [collectionName, url, collectionDescription, image, collectionAdded]);
 
   if (loading) {
     return (
